Move request test clients into cross_web.testing
