+++ src/cross_web/testing/__init__.py
@@ -1,0 +1,7 @@
+from .clients.base import FormRequestResult, JSONRequestResult, RequestClient
+
+__all__ = [
+    "FormRequestResult",
+    "JSONRequestResult",
+    "RequestClient",
+]
